Support for wide font names in TTF files set with the /fontfile switch
In OptionsClass.cpp the CSettings::GetFontNameFromFile_TTF method reads the font name into a char[] based on the ttRecord.uStringOffset and ttRecord.uStringLength, where the uStringLength is the size of the string in bytes. The
TTF spec show:

"Note that OS/2 and Windows both require that all name strings be defined in Unicode. Thus all 'name' table strings for platform ID = 3 (Windows) will require two bytes per character. Macintosh fonts require single byte strings."

( https://www.microsoft.com/typography/otspec/name.htm )

This means that the uStringLength can not be used as an array index into the name of the font since we are dealing with double byte characters.

So using something like:

ConEmu  /FontDir ".\\fonts" /Font "Roboto Mono Medium"

fails because the name can not be matched and ConEmu respons with

---------------------------
ConEmu 150913 [64]
---------------------------
Can't retrieve font family from file:
RobotoMono-Medium.ttf
Continue?
---------------------------
OK   Cancel
---------------------------

The following changes in this commit read the field as and array of wchar_t and address the issue.