Nuget package for Nuget.org — for referencing the stuff from Visual Studio projects. Unlike Choco's, this one has to carry all of the payload files inside and not on an URL.
The nuget with conemu WinForms control references this nuget, so just adding a nuget to your project sets it up for using the conemu control out of the box.
Here's the nuspec for Nuget (based on choco's) and the only file besides conemu binaries to be included with the package.