gh-673: GuiMacro SetWindowPosSize: fix changing size w/o changing pos (by @jcsullins).
SetWindowPosSize has logic that allows changing size or pos without
changing the other. However, the function (IntFromString) that parses
the pos args does not properly return if an int was parsed.
Thus, attempting to set size with pos unchanged (by passing a non
integer for the pos args) will always result is a pos of 0,0.

Fix IntFromString to properly return if an int was parsed.