Add a new option to auto-trim single lines.
When pasting a single line ending in newline, the user is presented with a
warning about pasting multi-line strings. This enables a new option to
automatically strip trailing newlines whenever we detect that a single line
string has been pasted. This prevents the warning and allows the user to more
conveniently edit single lines pasted into the console.