gh-2243: Fix arrow key issue in cygwin pseudo console.
If windows native app (such as vim, gnuplot, etc.) is started in
pseudo console in cygwin, arrow keys do not work. The cause is as
follows.

Cygwin 3.1.7, console handler uses xterm mode enabled. As a result,
arrow keys are translated to escape sequences such as "ESC[A",
"ESC[B", etc. When script command or cygwin-connector is started,
pty is opened. If windows native app is executed in pty, pseudo
console is activated. Pseudo console reconstructs escape sequences
"ESC[A" etc. to an arrow key event. However, if the escape sequence
is divided into multiple read, pseudo console fails to reconstruct
them.

This patch fixes the isseu.